pyproject.toml
[project]
name = "vision-rag-summarizer"
version = "0.1.0"
description = ""
authors = [
    {name = "sabaoon bedar", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pdf2image (>=1.17.0,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "langchain (>=0.1.14,<0.2.0)",
    "chromadb (>=0.4.24,<0.5.0)",
    "torch (>=2.2.0,<3.0.0)",
    "torchvision (>=0.17.0,<1.0.0)",
    "transformers (>=4.39.3,<5.0.0)",
    "accelerate (>=0.28.0,<1.0.0)",
    "opencv-python (>=4.9.0.80,<5.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "tts (>=0.22.0,<0.23.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "blobfile (>=3.0.0,<4.0.0)",
    "sentencepiece (>=0.1.99,<0.2.0)",
    "huggingface-hub (>=0.31.1,<0.32.0)",
    "pillow (>=9.2.0,<11.0.0)",
    "moviepy (>=2.1.2,<3.0.0)",
    "pyttsx3 (>=2.98,<3.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "imageio-ffmpeg (>=0.6.0,<0.7.0)",
]

[tool.poetry]
packages = [{include = "vision_rag_summarizer", from = "src"}]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
